Small refactoring (#525)
* Added refresh token request to postman collection

* refactor: update InternalServeException namespace

* refactor: add cancellation token to mail service
