+++ src/Core/Application/Common/Mailing/IMailService.cs
@@ -2,5 +2,5 @@ namespace FSH.WebApi.Application.Common.Mailing;
 
 public interface IMailService : ITransientService
 {
-    Task SendAsync(MailRequest request);
+    Task SendAsync(MailRequest request, CancellationToken ct);
 }
