+++ src/Infrastructure/Identity/RoleService.cs
@@ -2,7 +2,6 @@ using Finbuckle.MultiTenant;
 using FSH.WebApi.Application.Common.Events;
 using FSH.WebApi.Application.Common.Exceptions;
 using FSH.WebApi.Application.Common.Interfaces;
-using FSH.WebApi.Application.Identity;
 using FSH.WebApi.Application.Identity.Roles;
 using FSH.WebApi.Domain.Identity;
 using FSH.WebApi.Infrastructure.Persistence.Context;
